delete unused files, move triviaMachine to src and refactor guard with assertEvent
